Prevent the display scrolling too far when code is folded.
When a section of code is to be folded, if the fold-head is displayed the code down to the fold-head retains its screen position and all is well. However when the fold-head is off-screen and the user clicks on a margin to fold that code, none of the function will be displayed; and for large code sections, the whole display may change.
With this commit, in that situation the code below the fold-bottom (usually) retains its position, and the fold-head will always be shown.

git-svn-id: https://codelite.svn.sourceforge.net/svnroot/codelite/trunk@5408 9da81c78-c036-0410-9e1f-a2b0375e4b5a