Make FindinFiles work reliably again in locales.
Unlocalised labels (e.g. 'Active Project') weren't being found when tested against their translations. This is likely to happen when localisation has just started, as the type of search is serialised by name. As a result, no files get searched, so few matches are found. So test both the translated and untranslated labels.

git-svn-id: https://codelite.svn.sourceforge.net/svnroot/codelite/trunk@5496 9da81c78-c036-0410-9e1f-a2b0375e4b5a