Call-graph: Become custom-makefile friendly
The previous code only worked for CodeLite-generated makefiles. In particular it looked in the wrong place for the gmon.out file.
Change the functions used to get project info to ones that seem to work in either situation; look also in the binary dir; and, failing either, ask the user.

git-svn-id: https://codelite.svn.sourceforge.net/svnroot/codelite/trunk@5594 9da81c78-c036-0410-9e1f-a2b0375e4b5a