Fix for parenthesis autocompletion.
If 'Autocomplete Braces' is set, when the user types '(' a matching ')' is auto-provided; but only if the next character is white-space. This worked as desired most of the time, but was suboptimal if the next char was an already-matched ')'; see http://sourceforge.net/tracker/index.php?func=detail&aid=3534877&group_id=202033&atid=979960 for an example. The code now checks for this.

git-svn-id: https://codelite.svn.sourceforge.net/svnroot/codelite/trunk@5603 9da81c78-c036-0410-9e1f-a2b0375e4b5a