Don't try to GoTo a non-existent hyperlink
'Quick code navigation' lets a (by default) Sh-leftclick take you to the declaration of any symbol under the pointer. Sh-leftclick is also extensively used to extend the selection. The current code broke this: if there was no symbol under the pointer, an arbitrary amount of code at the beginning of the file got selected and scrolled to. Check for a hyperlink first, and allow the original behaviour if one doesn't exist.