If there's a name-clash wWhen adding a file to a project, name the clashing file
In !MSW, if a file 'foobar' is already present when you try to add FooBar, a messagebox is fired to tell you there's a potential clash. However this didn't reveal the name of the offending file, and since multiple different files can be added at once this caused a problem. The name is now shown in the messagebox.

Also, if the user chooses to cancel, abort the whole process not just this file.