Don't use "wxArrayFoo" when pretty-printing in gdb
I'd thought the return string from wxArrayFooPrinter.to_string() was unused, so just threw in "wxArrayFoo". In fact the string can be user-visible e.g. for a const wxArrayInt&, so use "wxArray<T>" instead. Ideal would be to specify the type, but I don't know how to do that inside a pretty-printer. CL displays the type in the next field anyway, so...