When reloading a file preserve any bookmarks
When a file changes outside the editor, usually because it's just changed on disc, CL offers to reload it. Doing so preserved any breakpoints, but bookmarks were discarded. Now bookmarks are reapplied too.

Of course if the file was considerably altered, they will be reapplied in the wrong places; but even that will usually be better than nothing.