Allow disabled breakpoints to stay disabled when the debugger stops
gdb (using the MI) can't cope with creating a disabled breakpoint, so when the debugger stops all breakpoints have been enabled. That's annoying when you have several of them that you want to be disabled for several gdb runs.

Instead let them stay disabled in CL and, once gdb restarts, explicitly tell gdb to disable them. NB Doing this used to cause gdb crashes in MSWin; but that was years ago so...