Really, really make loading disabled breakpoints work reliably
The previous way, 297c8fe, would have worked had I remembered to interrupt the debugger when trying to disable the breakpoints :/
The current way is easier and probably safer though: it catches the wxEVT_DEBUG_EDITOR_LOST_CONTROL event sent when control is lost i.e. when the debugger announces that it's running, and disables the BPs in the handler.