When saving and reloading a project file, try to reapply its 'fold' state
At present, and bookmarks and breakpoints are stored, then reapplied when a file is reloaded. Do the same for folds too.

I've not (yet) made this too intelligent. It works for unaltered files, and for altered ones providing the folds retain their old line-number. However it will fail downstream of any added/deleted lines. Which is still a lot better than nothing!