Multiple bookmark types: Create the underlying mechanism and use it for 'Find' bookmarks
The plan is to have two classes of bookmark: standard and 'find'. The latter will be used for the quickfind bar and the Find/Replace dialog.
Apart from other benefits, this will mean that setting 'Find' bookmarks doesn't clobber any current ones in the file, that being the current situation.
It also allows for several different flavours of 'standard' bookmark.