When CodeLite loads, update the lexer files if needed
This is a fix for https://sourceforge.net/p/codelite/bugs/931/
Before, when the CodeLite lexer files were updated with a new property or lexer, the user could (if he noticed!) choose to use the new file, thus losing any altered colour or fontsize preferences. The only way to avoid this would have been to alter the files by hand.
Now when it loads, CodeLite reconciles each lexer file with the equivalent installed one. This adds (for me) about 350ms to the starting time when changes are needed, otherwise just 50ms.