Partially fix a scrollbar bug when using the non-native notebook on wxGTK
Since the change to using the wxAuiNotebook by default, there's been an annoying glitch with scrollbar drawing: after the notebook selection is changed, any scrollbars are briefly drawn but then disappear, leaving a grey gap.
This commit fixes that when the selection changes because of a tab-click, but not when the change comes from a Workspace View selection.