Turn off calltips during Drag'n'Drop
One nasty habit of scintilla's calltips is firing during DnD, with the tip sometimes obscuring the desired drop zone. Calltips are now turned off while dragging.

For future reference, turning them on again afterwards was harder. wxSTC fires an event when drag starts, and another for a successful drop, but none for a failed one. So I've used wxEVT_LEFT_UP to catch the latter.