Make a (possibly futile) attempt to abort ectopic calltips
The nastiest habit of scintilla's calltips is firing just after you have left an editor. The tip then displays on top of any other program you are trying to use, and can't be cancelled until you return to the editor that owns the tip.
Try to prevent this, or at least reduce its frequency, by cancelling existing calltips when losing focus or leaving an editor; and, probably more usefully, by preventing them from firing unless scintilla has focus.

This commit is effectively untested as the problem happens randomly, presumably when the switch to a different program occurs just as the tip was about to fire. But it does no harm, and may help.