Two focus fixes for when activating an editor
Generally when an editor is selected e.g. when the user clicks on a tab, it is also focused. However this didn't happen when an editor was changed using the Workspace View 'tabs' pane, and it didn't happen when double-clicking on a cscope result. The latter was particularly annoying, as for some reason Undo/Redo UpdateUI considered it _was_ focused, yet clicking on the tool did nothing.

Fixed by introducing a new method, LiteEditor::DelayedSetActive, which is also available to plugins (i.e. cscope) via IEditor.