Use CallAfter() in the 'Goto Implementation/Declaration' implementation
The (ctags) implementation previously called LEditor::FindAndSelect synchronously, and that usually worked. Unfortunately, if the destination file wasn't currently loaded, there was a race situation between the load completing and the EnsureVisible() call which, at least for me on gtk, EnsureVisible() has started to lose.
So use CallAfter() to make it likelier that the file is loaded. That required a FindAndSelect()-type function returning void, which CallAfter() requires.