Add an alternative LEditor::FindAndSelect that returns void, and is implemented asynchronously
In wxGTK-3.1, perhaps because of its scintilla upgrade to 3.3.9, calls to FindAndSelect() fail to scroll to the correct location if the editor has been loaded just prior to the call. Presumably there's a race condition, with the scroll occurring too soon.
Fix this by substituting calls to FindAndSelectV(), which returns void (none of the relevant callers actually used the bool returned from the standard FindAndSelect()). This uses CallAfter() to create a delay before the real implementation.