It's now possible to tell cppcheck to test only certain 'configurations' (e.g. only branches where FOO is defined)
See http://codelite.org/forum/viewtopic.php?f=13&t=2539 for more details. Note that this is only available when cppcheck is run via a Workspace-Pane project context-menu, not from a workspace or file one, nor from the Explorer pane.

Definitions and undefines to be passed to cppcheck can be added to a new page of the settings dialog, which again is only available when accessed via a Workspace-Pane project context-menu.