Make the 'new keyboard shortcut' tooltip more helpful
The dialog lets you set a shortcut with or without modifiers (e.g. Shift). Any modifiers are set using checkboxes; however other programs expect the user to press their keys as well as the shortcut key. In CL this results in shortcuts like 'Shift-Shift-F'.
That's still possible, but at least the tooltip explains how to do it properly.