Hacked the two codelite-clang-format binaries to make them run on rpm-based distros too
The binaries were built on debian wheezy (and run on ubuntu too). Being in the llvm family, the inherit the customary dependency on libedit.so.2, which doesn't exist in fedora/openSUSE. However libedit.so _is_ universally available, so I've used ghex to edit the binaries to depend on that instead. This seems to work universally.

The downside is that the libedit development package becomes a _runtime_ dependency. However it already is for liblldb, for the same reason.