Make cscope GoTo work correctly again for folded code
The 68ab6e9c commit added a SetLineVisible() call to LEditor::DoFindAndSelect. This uses SetFirstVisibleLine(lineno), which works on lines visible on the screen, not doc lines. As a result, the 'visible' line was well off the bottom of the sceen if much code was folded.

Fix this with a call to VisibleFromDocLine().