Link LLDBDebugger to LLDBProtocol before liblldb
When linking with -Wl,--as-needed the linker removes libraries which
are unreferenced in the previous files specified on the command line.
If LLDBProtocol is specified after liblldb, the linker will have
already removed liblldb as unrefernced before LLDBProtocol was linked in.
To prevent this, link LLDBProtocol in before linking liblldb in.