Provide some minimal customisation of xterm on Linux
The CodeLite debugger terminal is now xterm, which on most Linux distros is decidedly minimal and unmodern. In particular it might not have a scrollbar, and by default scroll-back is only 64 lines.
So force a scrollbar to appear and set scroll-back to 1000; any computer capable of running CodeLite should have enough memory for that. More customisation can be done by providing a ~/.Xresources file: see http://codelite.org/LiteEditor/Linux-specificTheXtermDebuggerTerminal