UnRedo: fix a situation where CLCommandProcessor could become desynchronised from the scintilla one
It was possible to get a situation where, after an Undo, the next potential Undo would be e.g. some typing. If the user now added more typing, scintilla was amalgamating this into the 'next potential Undo' but CLCommandProcessor wasn't. As a result the menu dropdown would be off-by-one for each time this situation had occurred.
Fixed by changing the scintilla behaviour.