When fonts and colours have been updated, make sure any session is loaded
Loading a session happens when a wxEVT_COLOURS_AND_FONTS_LOADED event is handled by clMainFrame. This event is normally sent by ColoursAndFontsManager::Load just before the frame is created, so the pending event arrives at the correct time. However the code-path when lexers are updated sends the event earlier, and it before the frame is ready; so send another afterwards.