Partial fix for a focus bug involving the QuickFind bar
Recent Highlight Matching Word changes mean that scintilla sets keyboard focus on the current editor after a delay. There is a consequent race situation between that and the user clicking on a QuickFind field; who wins depends on the editor's line-count and the visibility of the word currently being matched. When QuickFind loses the race, it never receives focus, which means that attempts by the innocent user to paste a string into it will result in the string being added to the editor; worse, that addition happens at the current insertion point, which will often be off the screen.

Work somewhat around the issue by increasing the delay used in QuickFindBar::OnReceivingFocus. This is imperfect, but considerably reduces the bug's incidence.