Make the multiarch-setting code work correctly on Linux with cmake >3.0.0
With cmake 2.8 the GNUInstallDirs module ran whatever the value of CMAKE_INSTALL_PREFIX. Since 3.0.1 it only runs if CMAKE_INSTALL_PREFIX is set to /usr. As the default value is /usr/local, that will seldom happen; as a result the default location for CodeLite libs changed to /usr/lib/codelite in debian jessie, which uses cmake 3.0.2.

As we document that the default prefix is /usr, explicitly set it to this for the offending cmake versions; but only if the user hasn't supplied his own selection.