Remove uses of __DATE__ and __TIME__ from libctags
There is an effort in Debian (and other distros) to be able to create
reproducible binaries. Including a timestamp embedded in the executable will
clearly make it slightly different each time it's built. After removing these
two instances from libctags, codelite can be built reproducibly (with an
experimental toolchain).

More info:
https://wiki.debian.org/ReproducibleBuilds/About

The 7.0+dfsg-2 upload to Debian had this patch applied and this is the log:
https://reproducible.debian.net/rb-pkg/unstable/amd64/codelite.html