Make it possible to use 'konsole' as the debugger terminal
In Linux the terminal has been hard-wired to be xterm, which may not have been (any)|(every)one's choice, but was known to work. The problem with others (==konsole) was due to them reporting the full path to the 'sleep' utility, not just "sleep".
Using and searching for that allows all the standard terminals to be detected.