When using konsole as the debugger terminal, work around its bug-features
There are two problems when using konsole as the debugger terminal. The first is that it doesn't automatically close when its only child, the 'sleep' program, is terminated, even if sigterm or sighup are used. Instead it shows a 'sleep has crashed' warning and hangs.
Fix this by terminating konsole too.

However this reveals the other bug-feature: by default, opening a second (or subsequent) konsole instance doesn't actually do so; instead it opens a new thread of the existing one. So a sigterm etc will kill them all.
Fix this by passing the --separate option when starting the terminal.