Use the new PromptForYesNoCancelDialogWithCheckbox() function to decide whether to save the Perspective on exiting
When exiting CodeLite the previous situation was to allow the user to save and exit or cancel the save.
A combination of 'cancel' and 'remember this answer' had the unfortunate consequence of never being able to exit.

The new way allows a choice of saving the perspective or not; but only this is 'remembered' so closing always remains possible.