git plugin: Added a 'blame' dialog
This displays the current editor's contents. Alongside each line is most-recent commit hash, author and date in the margin.
Hovering over the margin for a (configurable) time will fire a popup showing that commit's log.
Double-clicking on a line will go back in history, configurably showing either the contents at the time of that line's commit or, more usefully, the commit just before.
It is also possible to write in extra options (e.g. -L100,130) that will be passed to git when the blame is refreshed.