Locate dependencies on unix via which
php, clang_format, php-cs-fixer/phpbcf, phpcs, phpmd are now configured
automatically (in the case of phar-files it only applies if on unix) if
installed on the system.

The autodetect is also moved to the loading of the config to avoid
having to restart CL for it to autodetect again if values are emptied,
giving the user more emediat feedback.