Prevent an intermittent wx assert when cancelling the Project Settings dialog
Code to save the Project Settings dialog's selected page was called from the ProjectSettingsDlg dtor. This resulted in a wxEVT_EDITOR_CONFIG_CHANGED event that, presumably because of a race condition, often resulted in the assert "nNew != dynamicEvents.size()" failed in SearchDynamicEventTable() when the dialog was cancelled, though not with the OK button.

This is fixed by moving the page-saving call to OnButtonOK() so it doesn't happen if the dialog is cancelled (which is probably the expected behaviour anyway).