Cscope: After clicking on an entry, try harder to centre the target line.
Clicking on an entry in the Output View's Cscope tab will open the target file, goto that line and should centre it on the page. However the centring often failed if the file needed to be loaded, presumably because the call to CenterLine() was made before the file was available. So repeat it, delayed.