GitCommit dialog: Show the previous commit message when amending, and make the History more useful
If the 'Amend the previous commit' box is checked, the title and body of the last commit message are now shown, for information and possible alteration.

If the 'History' button is clicked, the resulting dialog now shows the (hash and) title of the last 100 commits to the currently active git repo; it had previously displayed the last few entries to any repo. Selecting an entry now inserts both its title and body into the commit box, not just the title.