When removing a file from a project fails, try again using its absolute path
A CodeLite workspace project stores its entries as relative filepaths, and the code that implements removal of an entry assumes that it is relative.
However this assumption will not always be true if the user had edited the project by hand, when he may have used absolute filepaths.
So, if removal fails, try again using the absolute path.