Diff panel Overview bar: Make it clearer which markers (if any) correspond to the currently-displayed lines
With a long file and many diffs it can be hard to know which Overview bar markers correspond to the lines visible in the editor. This is especially true in horizontal mode, where the scrollbar thumb does not help.

The bar now has a 'viewport' marker that highlights the part of the bar that matches the displayed lines.