Git plugin: Store the last-used path for each project for future reuse
When a workspace is loaded, the previous code looked for any git repo first in the active project dir, then in its parent (which is probably the workspace dir). However this failed to cope with a repo path that was outside the workspace: see https://github.com/eranif/codelite/issues/1358.

Each projects' last repo path is now saved, and is used if no git directory is found in the project or workspace. This should cope with most sane setups, though it will fail if e.g. the workspace dir contains a different project's git.