When running the Setup wizard requires a restart, don't ask the user whether to save their Perspective
If a new user changed the Developer Profile when running the Setup wizard, the 'Save your Perspective?' annoying dialog would fire. This was not only unnecessary (he wouldn't have had a chance to change Perspective yet) but also confusing as the dialog would get lost behind the new CodeLite instance before he had a chance to read it.