In Linux, try even harder to cope with project files that are symlinks
Previous fixes had made it possible correctly to match a file to a Project filepath where either both, or just the file, were actually symlinks.
However it still failed if only the Project filepath was a symlink. That caused problems when:
* a Project was in a dir whose path contained a symlink
* the project was being run in gdb; a breakpoint was hit; then a frame for a currently-unloaded file was loaded from the callstack
That file would be called by its real path, and the editor would be identified as such.
One result (I expect there are others) would be that a breakpoint added to that mislabelled editor could not be deleted later when the 'real' editor was loaded.