Don't try directly to set a conditional watchpoint; gdb can't cope
There isn't a GDB/MI command that creates a conditional watchpoint, and pretending it's a breakpoint (as did the old code) fails altogether.
Instead create a normal watchpoint, and later edit it to add the condition.
Disable the 'Condition' box if a watchpoint is being created, and use its tooltip explain why.