Fix phpstan intergration
The the parameter name appears to have been changed.
Path are relative to CWD because phpstan is special.
Drop input for phpstan config as this was fully implemented.
Assume that phpstan.neon exists for the current project root (this is
required because else the output will contain a notice that breaks the
XML), note the notice is outputted on stderr but we still include it in
the output buffer?