Showing the parent folder on each tab is now optional
Since 8c5220d17 tabs are labelled with the parent folder in addition to the filename.
This commit makes that optional.

Note that updating the setting has to be applied from the wxEVT_EDITOR_SETTINGS_CHANGED handler;
doing it direct from EditorSettingsDockingWindows::Save will fail as its 'options' object
has not yet been serialised.