Remove --use-temp-file from wx-config's rcflags
wx-config always adds "--use-temp-file" to its rcflags to be
used as a windres parameter when compiling MS Windows resources.

However, this parameter is a relic from the difficult Windows 9x
times and is known to break the compilation in latest GCC versions.

wxWidgets does not use this parameter for its build anymore,
stop using it in wx-config as well.