Update the build dependencies for linux
Currently in Debian/Ubuntu wxWidgets libraries are outdated for the needs of the project (See https://github.com/eranif/codelite/issues/2674#issuecomment-780140511). This includes a better description of what to do to get the proper versions. Also, my test conclude that the minimal installable packages to get it working with the proposed repository are (with their own dependencies):

- libwxbase3.1-dev
- libwxbase3.1unofficial3-dev
- libwxgtk3.1unofficial3-dev