Git Plugin: Add an extra way to find a project's .git dir.
A project's .git dir might occasionally be in an located outside the project, even in a different partition. For that situation, there is now a new field in the Git Settings dialog that allow users directly to enter its path.

Also, when searching for .git/ in or above the project dir, avoid symlink issues by using CLRealPath.