Improve editor symlink handling
This patch improves CodeLite with projects that have symlinks in the path

If you open a file in a project that has symlinks in its path - you'll see the expanded/resolved path of the file in the window title.
Also you'll typically see that CodeLite is unable find the file in the Workspace View

There are several problems in respect to symlinks and the use of realpath() calls that this patch addresses:

 - Clangd LSP returns realpath() paths

 - mainbook uses realpath() extensively to figure out if a file is already opened (in another tab).

 - CodeLite Find-In-Files uses realpath() extensively (and in the end - clicking on the found file opens it)

 - etc...

This patch basically fakes the use of FileUtils::RealPath() calls - on most occasions just returning the unmodified path.

An optional argument to FileUtils::RealPath() (defaults to false) can force the realpath() conversion (this is only used in very few occasions in this patch)