Update Assistant Submodule
The assistant submodule has been updated to incorporate the
latest changes from its repository. Mainly, an incorrect "Tools" JSON
was provided for functions that do not require parameters.

Added a new task to the TODO list for providing a "codelite-api.lua" file
in the installation folder to enable code completion support for Lua LSPs.
This will improve the developer experience when working with Lua scripts
in the IDE.

** Generated by CodeLite. **

Signed-off-by: Eran Ifrah <eran@codelite.org>